# backend/src/test/resources/openapi/phase2-openapi-baseline.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "知家 API",
    "description" : "知家家庭物品管理系统 REST API",
    "version" : "1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/audit-logs" : {
      "get" : {
        "tags" : [ "audit-log-controller" ],
        "operationId" : "auditLogs",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "action",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "actorAccountId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "outcome",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuditLogPageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "AccountInfo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "username" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "AuditLogItemResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "action" : {
            "type" : "string"
          },
          "outcome" : {
            "type" : "string"
          },
          "actorAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "actor" : {
            "$ref" : "#/components/schemas/AccountInfo"
          },
          "subjectAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "subject" : {
            "$ref" : "#/components/schemas/AccountInfo"
          },
          "detail" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "ipAddress" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AuditLogPageResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditLogItemResponse"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}